+++ includes/class-settings.php
@@ -98,8 +98,6 @@ class SUPER_Settings {
         }
         $statuses = $new_statuses;
         unset($new_statuses);
-
-        $settings = stripslashes_deep( $settings );
         
         $submission_count = 0;
         if( (isset($settings['id'])) && ($settings['id']!=0) ) {
@@ -143,7 +141,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'header_from_type' => array(
                     'name'=> __( 'Send email from:', 'super-forms' ),
@@ -166,7 +164,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'header_from_type',
                     'filter_value'=>'custom',
-                    'allow_empty'=>true,
+                    
                 ),
                 'header_from_name' => array(
                     'name' => __( 'From name:', 'super-forms' ),
@@ -176,7 +174,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'header_from_type',
                     'filter_value'=>'custom',
-                    'allow_empty'=>true,
+                    
                 ),
 
                 // @since 2.8.0 - custom reply to headers
@@ -198,7 +196,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'header_reply_enabled',
                     'filter_value'=>'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'header_reply_name' => array(
                     'name' => __( 'Reply to name:', 'super-forms' ),
@@ -208,7 +206,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'header_reply_enabled',
                     'filter_value'=>'true',
-                    'allow_empty'=>true,
+                    
                 ),
 
                 'header_subject' => array(
@@ -219,7 +217,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'email_body_open' => array(
                     'name' => __( 'Body header:', 'super-forms' ),
@@ -229,7 +227,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'email_body' => array(
                     'name' => __( 'Body content:', 'super-forms' ),
@@ -239,7 +237,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'email_body_close' => array(
                     'name' => __( 'Body footer:', 'super-forms' ),
@@ -249,7 +247,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'email_loop' => array(
                     'name' => __( 'Field Loop:', 'super-forms' ),
@@ -260,7 +258,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true, 
+                     
                 ),
                 // @since 3.1.0 - auto line breaks
                 'email_body_nl2br' => array(
@@ -274,7 +272,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'send',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'header_cc' => array(
                     'name' => __( 'CC:', 'super-forms' ),
@@ -335,7 +333,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,  
+                      
                 ),
                 'confirm_from_type' => array(
                     'name'=> __( 'Send email from:', 'super-forms' ),
@@ -358,7 +356,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm_from_type',
                     'filter_value'=>'custom',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_from_name' => array(
                     'name' => __( 'From name:', 'super-forms' ),
@@ -368,7 +366,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm_from_type',
                     'filter_value'=>'custom',
-                    'allow_empty'=>true,
+                    
                 ),
 
                 // @since 2.8.0 - custom reply to headers
@@ -390,7 +388,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm_header_reply_enabled',
                     'filter_value'=>'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_header_reply_name' => array(
                     'name' => __( 'Reply to name:', 'super-forms' ),
@@ -400,7 +398,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm_header_reply_enabled',
                     'filter_value'=>'true',
-                    'allow_empty'=>true,
+                    
                 ),
 
                 'confirm_subject' => array(
@@ -410,7 +408,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_body_open' => array(
                     'name' => __( 'Body header:', 'super-forms' ),
@@ -420,7 +418,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_body' => array(
                     'name' => __( 'Body content:', 'super-forms' ),
@@ -430,7 +428,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_body_close' => array(
                     'name' => __( 'Body footer:', 'super-forms' ),
@@ -440,7 +438,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 // @since 3.1.0 - auto line breaks
                 'confirm_body_nl2br' => array(
@@ -454,7 +452,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'confirm',
                     'filter_value'=>'yes',
-                    'allow_empty'=>true,
+                    
                 ),
                 'confirm_header_cc' => array(
                     'name' => __( 'CC:', 'super-forms' ),
@@ -612,7 +610,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'conditionally_save_entry',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
 
 
@@ -646,7 +644,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent'=>'enable_custom_entry_title',
                     'filter_value'=>'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'contact_entry_add_id' => array(
                     'default' => self::get_value( $default, 'contact_entry_add_id', $settings, '' ),
@@ -729,7 +727,7 @@ class SUPER_Settings {
                     'values' => array(
                         'true' => __( 'Show thank you message', 'super-forms' ),
                     ),
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_thanks_title' => array(
                     'name' => __( 'Thanks Title', 'super-forms' ),
@@ -738,7 +736,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_show_thanks_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_thanks_description' => array(
                     'name' => __( 'Thanks Description', 'super-forms' ),
@@ -748,7 +746,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_show_thanks_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_preload' => array(
                     'name' => __( 'Preloader (form loading icon)', 'super-forms' ),
@@ -948,7 +946,7 @@ class SUPER_Settings {
                     'values' => array(
                         'true' => __( 'Hide form after submitting', 'super-forms' ),
                     ),
-                    'allow_empty'=>true,
+                    
                 ),
                 // @since 2.0.0  - reset / clear the form after submitting
                 'form_clear_after_submitting' => array(
@@ -1005,7 +1003,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_locker',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_locker_msg_title' => array(
                     'name' => __( 'Lock message title', 'super-forms' ),
@@ -1013,7 +1011,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_locker_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_locker_msg_desc' => array(
                     'name' => __( 'Lock message description', 'super-forms' ),
@@ -1022,7 +1020,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_locker_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_locker_hide' => array(
                     'default' => self::get_value( $default, 'form_locker_hide', $settings, 'true' ),
@@ -1033,7 +1031,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'form_locker',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'form_locker_reset' => array(
                     'name' => __( 'Select when to reset the form lock', 'super-forms' ),
@@ -1104,7 +1102,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'user_form_locker',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'user_form_locker_msg_title' => array(
                     'name' => __( 'Lock message title', 'super-forms' ),
@@ -1112,7 +1110,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'user_form_locker_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'user_form_locker_msg_desc' => array(
                     'name' => __( 'Lock message description', 'super-forms' ),
@@ -1121,7 +1119,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'user_form_locker_msg',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'user_form_locker_hide' => array(
                     'default' => self::get_value( $default, 'user_form_locker_hide', $settings, 'true' ),
@@ -1132,7 +1130,7 @@ class SUPER_Settings {
                     'filter'=>true,
                     'parent' => 'user_form_locker',
                     'filter_value' => 'true',
-                    'allow_empty'=>true,
+                    
                 ),
                 'user_form_locker_reset' => array(
                     'name' => __( 'Select when to reset the form lock', 'super-forms' ),
@@ -1531,7 +1529,7 @@ class SUPER_Settings {
                     'values' => array(
                         'true' => __( 'Show progress bar for Multi-part', 'super-forms' ),
                     ),
-                    'allow_empty'=>true,
+                    
                 ),
 
                 'theme_progress_bar_colors' => array(
@@ -1565,7 +1563,7 @@ class SUPER_Settings {
                     'values' => array(
                         'true' => __( 'Show steps for Multi-part', 'super-forms' ),
                     ),
-                    'allow_empty'=>true,
+                    
                 ),
 
                 'theme_progress_step_colors' => array(
@@ -1755,7 +1753,7 @@ class SUPER_Settings {
                     'name' => __( 'Global font family', 'super-forms' ),
                     'label' => __( '(leave blank for default) e.g: \'Raleway\', sans-serif', 'super-forms' ),
                     'default' => self::get_value( $default, 'font_global_family', $settings, '"Open Sans",sans-serif' ),
-                    'allow_empty'=>true,
+                    
                 ),
             )
         );
@@ -1794,13 +1792,13 @@ class SUPER_Settings {
                 'form_button' => array(
                     'name' => __('Button name', 'super-forms' ),
                     'default' => self::get_value( $default, 'form_button', $settings, __( 'Submit', 'super-forms' ) ),
-                    'allow_empty'=>true,
+                    
                 ),
                 // @since 2.0.0
                 'form_button_loading' => array(
                     'name' => __('Button loading name', 'super-forms' ),
                     'default' => self::get_value( $default, 'form_button_loading', $settings, __( 'Loading...', 'super-forms' ) ),
-                    'allow_empty'=>true,
+                    
                 ),
 
                 'theme_button_colors' => array(
@@ -1943,7 +1941,7 @@ class SUPER_Settings {
                     'desc' => __('Put each on a new line.<br />Example:<br />fieldname|Field label<br />email|Email<br />phonenumber|Phonenumber', 'super-forms' ),
                     'default' => self::get_value( $default, 'backend_contact_entry_list_fields', $settings, "email|Email\nphonenumber|Phonenumber\nmessage|Message" ),
                     'type' => 'textarea', 
-                    'allow_empty'=>true,
+                    
                 ),
 
                 // @since 3.4.0 - contact entry status
@@ -1952,7 +1950,7 @@ class SUPER_Settings {
                     'desc' => __('Put each on a new line.<br /><br />Format:<br />name|label|bg_color|font_color<br /><br />Example:<br />pending|Pending|#808080|#FFFFFF<br />processing|Processing|#808080|#FFFFFF<br />on_hold|On hold|#FF7700|#FFFFFF<br />accepted|Accepted|#2BC300|#FFFFFF<br />completed|Completed|#2BC300|#FFFFFF<br />cancelled|Cancelled|#E40000|#FFFFFF<br />declined|Declined|#E40000|#FFFFFF<br />refunded|Refunded|#000000|#FFFFFF', 'super-forms' ),
                     'default' => self::get_value( $default, 'backend_contact_entry_status', $settings, $backend_contact_entry_status ),
                     'type' => 'textarea', 
-                    'allow_empty'=>true,
+                    
                 ),
 
                 // @since 1.2.9
@@ -1963,7 +1961,7 @@ class SUPER_Settings {
                         'true' => __('Add the form name to the contact entry list', 'super-forms' ),
                     ),
                     'type' => 'checkbox',
-                    'allow_empty'=>true,
+                    
                 ),
 
                 // @since 3.1.0 - allow to display IP address to the contact entry column
@@ -1999,6 +1997,26 @@ class SUPER_Settings {
             ),
         );
         $array = apply_filters( 'super_settings_after_custom_css_filter', $array, array( 'settings'=>$settings ) );
+
+
+        /** 
+         *  Custom JS
+         *
+         *  @since      4.2.0
+        */
+        $array['custom_js'] = array(        
+            'hidden' => true,
+            'name' => __( 'Custom JS', 'super-forms' ),
+            'label' => __('Add custom JavaScript code', 'super-forms' ),
+            'fields' => array(
+                'theme_custom_js' => array(
+                    'name' => __('Custom JS', 'super-forms' ),
+                    'default' => self::get_value( $default, 'theme_custom_js', $settings, '' ),
+                    'type' => 'textarea', 
+                ),
+            ),
+        );
+        $array = apply_filters( 'super_settings_after_custom_js_filter', $array, array( 'settings'=>$settings ) );
 
         
         /** 
@@ -2363,21 +2381,6 @@ class SUPER_Settings {
         }else{
             return ( !isset( $settings[$name] ) ? $default : $settings[$name] );
         }
-    }
-    
-    /**
-     * Reformat the settings
-     * @param  array $settings
-     *
-     *	@since		1.0.0
-    */
-    public static function format_settings( $settings=null ) {
-        if($settings!=false){
-            foreach($settings as $k=>$v) {
-                $settings[$k] = stripslashes($v);
-            }
-        }
-        return $settings;
     }
     
 }